requirements.txt
discord.py>=2.4.0
SQLAlchemy>=2.0.0
alembic>=1.11.1
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
matplotlib>=3.7.0
requests>=2.31.0
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
redis>=5.0

dateparser
